requirements.txt
fastapi
tortoise-orm
pydantic<2.0.0
uvicorn[standart]
python-dotenv
asyncpg